+++ portico_webworks_plugin.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: Portico Webworks Hotel Website Manager
  * Description: Portico Webworks plugin.
- * Version: 0.7.7
+ * Version: 0.7.8
  * Author: Portico Webworks
  * Author URI: https://porticowebworks.com
  * License: Proprietary (All Rights Reserved) - Unauthorized use by other companies or for any purposes is prohibited without written permission.
@@ -13,7 +13,7 @@ if (!defined('ABSPATH')) {
 }
 
 define('PW_PLUGIN_FILE', __FILE__);
-define('PW_VERSION', '0.7.7');
+define('PW_VERSION', '0.7.8');
 
 function pw_apply_install_defaults() {
 	if (get_option('pw_install_defaults_applied', 0)) {
@@ -27,7 +27,12 @@ function pw_apply_install_defaults() {
 	update_option( 'pw_seed_taxonomies', 1 );
 }
 
-register_activation_hook(PW_PLUGIN_FILE, 'pw_apply_install_defaults');
+function pw_plugin_activation() {
+	pw_apply_install_defaults();
+	set_transient('pw_activation_settings_notice', 1, 300);
+}
+
+register_activation_hook(PW_PLUGIN_FILE, 'pw_plugin_activation');
 
 add_action( 'init', function() {
 	if ( ! get_option( 'pw_seed_taxonomies', 0 ) ) {